CSS styling and logging system have been added.

+++ php/main.php
@@ -6,7 +6,12 @@
   $app->get('/', function() {
     require_once 'config.php';
     require_once 'functions.php';
-    $connect=connectDB();
+    $shortener=new Shortener;
+    $connect=$shortener->connectDB();
+    //check for DB connection
+    if ($connect->connect_errno) {
+      return "There is no connection to the DB";
+    }
     $lock=$connect->query("LOCK TABLES `urls` WRITE");
     //delete URL pairs which are older than 15 days
     $success=$connect->query("DELETE FROM `urls` WHERE DATEDIFF(CURDATE(), `date`)>=15");
@@ -17,23 +22,28 @@
     //return file_get_contents('http://DOMAIN/php/getshorturl.php?longurl='.urlencode('URL'));
   });
   $app->get('/{url}', function($url) {
-    //check for correct simbols
-    if(!preg_match(EXPRESSION, $url)) {
-	     die('That is not a valid short URL');
-    }
     require_once 'config.php';
     require_once 'functions.php';
-    $shorturl="http://".$_SERVER["HTTP_HOST"]."/".implementFilters($url);
-    $connect=connectDB();
+    $shortener=new Shortener;
+    $connect=$shortener->connectDB();
+    //check for DB connection
+    if ($connect->connect_errno) {
+      die ("There is no connection to the DB");
+    }
+    //check for correct simbols
+    if(!preg_match(EXPRESSION, $url)) {
+      die ("That is not a valid short URL");
+    }
+    $shorturl="http://".$_SERVER["HTTP_HOST"]."/".$shortener->implementFilters($url);
     //check if there is such URL in the DB
-    $id=mysqli_fetch_row(verifyShortURL($shorturl, $connect));
+    $id=mysqli_fetch_row($shortener->verifyShortURL($shorturl, $connect));
     if (!$id) {
-      die("There is no such short URL in the DB");
+      die ("There is no such short URL in the DB");
     }
     //get initial URL
     else {
-      $longurl=mysqli_fetch_row(getLongURL($id[0], $connect));
-      incrementUsage($id[0], $connect);
+      $longurl=mysqli_fetch_row($shortener->getLongURL($id[0], $connect));
+      $shortener->incrementUsage($id[0], $connect);
     }
     $connect->close();
     //redirect to initial URL
